update show category list

+++ src/main/java/com/example/pbl3_1/dao/GenericDAO.java
@@ -8,5 +8,5 @@ public interface GenericDAO<T> {
     <T> List<T> query(String sql, RowMapper<T> rowMapper, Object... parameters);
     void update (String sql, Object... parameters);
     Long save (String sql, Object... parameters);
-    int count(String sql, Object... parameters);
+    Long count(String sql, Object... parameters);
 }
